power: qpnp-smb2: Remove useless "qcom,thermal-mitigation-china" node
Actually, the addition of the node "qcom,thermal-mitigation-china"
absolutely useless here as the current threshold set to dtsi is the
same as the node "qcom,thermal-mitigation", then added a hw check to
differentiate the Global and Indian versions to this thermal node. So
remove hw check for node "qcom,thermal-mitigation" and also delete node
"qcom,thermal-mitigation-china". So this allows the Indian or Global
version to use node "qcom,thermal mitigation" by default.

Signed-off-by: Fiqri Ardyansyah <fiqri15072019@gmail.com>